refactor: split metrics into separate files
